requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
openpyxl>=3.1.0
matplotlib>=3.7.0

# Optional dependencies (app works without these but with reduced functionality)
tableone>=0.8.0
lifelines>=0.27.0
python-docx>=0.8.11
